Refactor robot status management and update imports
- Moved robot-related enums (RobotMode, SafetyStatus, ProgramState) to a new module `robot_enums.py` for better organization.
- Updated `robot_status_monitor.py` to utilize the new enums and removed redundant enum definitions.
- Adjusted `global_vars.py` to import enums from `robot_enums.py`.
- Enhanced the Python import helper with updated paths for improved code completion related to the new enum structure.
